refactor: スポンサープランのファイル名を変更
